Reorganized the files, moved the old main into an old_main file and moved the new_main into just main. Also got packet receiving working on the main app not just the listener_test.
